+++ demo/src/App.tsx
@@ -1,11 +1,11 @@
-// import { DemoScene } from './scenes/demo'
+import { DemoScene } from './scenes/demo'
 import Nyan from './scenes/nyan/nyan'
 
 function App() {
   return (
     <>
       <Nyan />
-      {/* <DemoScene /> */}
+      <DemoScene />
     </>
   )
 }